Add locgic to aco_wc_initialize_payment function so a redirect to thankyou page is done if Avarda status already is Completed
This stops issues if redirect back to the store (for card and Swish payments) happens after the session expiration time. This also means that payment methods that require a redirect to a external payment page will redirect the customer directly to the thank you page after completed payment. Previously the customer was redirected to the checkout page and we waited/listened for the completedPurchaseCallback JS callback.