Change to fake the customer session
Needs to be done to prevent the API from being called to create a session before the frontend has gotten the address change sent to it. Otherwise the solution would not work properly for logged out users when shipping zones are used.

This does not actually modify the customers session at all, only loads in the users session data and sets it for the API call to use to calculate the shipping that should be available for the user with the same address, only setting the postcode and country for the actual calculation.

This has been tested with coupons. limited coupons and free shipping coupons to ensure they work properly.