Change to aco_wc_after_wrapper to insert hidden jwt token field
This is done to prevent issues with plugins or themes triggering the `woocommerce_checkout_fields` before the session from Avarda is created, which would cause the JWT error on each page load, since the token would never be added to the input field.

The session is currently created the first time on the action `aco_wc_before_checkout_form` and this will ensure that the field is always added after that.