Prevent init from running twice
Prevent init from running twice, which seemingly happens in some instances (always resolved on reload).

Fixes the merchant reported issue of double click events being triggered on pickup point select, which creates instant closing of the select after attempt to open.

Could this be prevented in a different way? The script is not included twice on the page.